backend/requirements.txt
# PantryPal Backend Requirements
# Python 3.11+

# ============================================
# Web Framework
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ============================================
# Database
# ============================================
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0  # Async PostgreSQL driver

# ============================================
# Authentication & Security
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.1

# ============================================
# Data Validation
# ============================================
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# ============================================
# HTTP Client (for external APIs)
# ============================================
httpx==0.25.1
requests==2.31.0

# ============================================
# Testing
# ============================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-env==1.1.3
faker==20.1.0  # Generate fake data for testing

# ============================================
# Code Quality
# ============================================
ruff==0.1.6  # Fast Python linter
black==23.11.0  # Code formatter
mypy==1.7.1  # Type checking

# ============================================
# Utilities
# ============================================
python-dateutil==2.8.2
pytz==2023.3
pillow==10.1.0  # Image processing (for future features)

# ============================================
# Development Tools
# ============================================
ipython==8.17.2
watchdog==3.0.0  # File watching for auto-reload

# ============================================
# Production
# ============================================
gunicorn==21.2.0  # Production server
sentry-sdk[fastapi]==1.38.0  # Error tracking (optional)

# ============================================
# API Documentation
# ============================================
# (FastAPI includes Swagger UI by default)
